refactor: move analysis creation to data layer
